valta/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
sqlalchemy>=2.0.20
python-dotenv>=1.0.0
httpx>=0.25.0
openai>=1.52.0
anthropic>=0.40.0
docling>=2.54.0
pypdf>=3.17.0
python-docx>=1.1.2
pandas>=2.1.0
numpy>=1.25.0